requirements.txt
discord.py
feedparser
watchdog